requirements.txt
streamlit==1.3.1
pandas==1.3.5
seaborn==0.11.2
plotly==5.1.0
numpy==1.21.2
matplotlib.pyplot==3.5.0